+++ Source/Core/Core/PowerPC/Gekko.h
@@ -803,5 +803,5 @@ enum
 	EXCEPTION_PERFORMANCE_MONITOR = 0x00000100
 };
 
-inline s32 SignExt16(s16 x) {return (s32)(s16)x;}
+inline s32 SignExt16(s16 x) {return (s32)x;}
 inline s32 SignExt26(u32 x) {return x & 0x2000000 ? (s32)(x | 0xFC000000) : (s32)(x);}
